move useCodeIntel hook to /enterprise/ tree
